DRAGONS: Rename VabSound.* to vabsound.*
