+++ engines/dragons/vabsound.h
